Move parser module into its own source file
